requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.28.0,<0.29
djangorestframework-simplejwt==5.3.1